Move fork_choice_rule.js to ./bft folder
