+++ framework/src/modules/chain/bft/fork_choice_rule.js
