+++ framework/test/jest/unit/specs/modules/chain/blocks/fork_choice_rule.spec.js
@@ -5,7 +5,7 @@ const {
 	isDuplicateBlock,
 	isTieBreak,
 	isValidBlock,
-} = require('../../../../../../../src/modules/chain/blocks/fork_choice_rule');
+} = require('../../../../../../../src/modules/chain/bft/fork_choice_rule');
 const { Slots } = require('../../../../../../../src/modules/chain/dpos');
 
 const EPOCH_TIME = new Date(Date.UTC(2016, 4, 24, 17, 0, 0, 0)).toISOString();
